# api/routes/plugin_routes.py
from fastapi import APIRouter, Depends
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from api.controllers.auth_controller import verify_token_controller
from api.controllers.plugin_controller import (
    export_all_outputs_controller,
    export_plugin_output_controller,
    get_available_outputs_controller,
    get_user_plugins_controller,
    get_plugin_by_id_controller,
    create_plugin_controller,
    update_plugin_controller,
    toggle_plugin_controller,
    delete_plugin_controller,
    execute_plugin_controller,
    update_installed_plugin_controller,
    get_latest_execution_controller,
    validate_plugin_controller,
)
from api.models.plugin import (
    PluginCreate,
    PluginExportAllRequest,
    PluginExportRequest,
    PluginUpdate,
    PluginToggle,
    PluginExecute,
    PluginValidate,
)
from typing import Annotated

router = APIRouter(prefix="/plugins", tags=["Plugins"])
bearer_scheme = HTTPBearer()


def get_current_user(
    credentials: HTTPAuthorizationCredentials = Depends(bearer_scheme),
):
    result = verify_token_controller(credentials.credentials)
    return result["user"]


CurrentUser = Annotated[dict, Depends(get_current_user)]


@router.post("/export", summary="Export a single plugin output")
def export_plugin_output(request: PluginExportRequest, current_user: CurrentUser):
    return export_plugin_output_controller(request, current_user["id"])


@router.post("/export/all", summary="Export all outputs from an execution")
def export_all_outputs(request: PluginExportAllRequest, current_user: CurrentUser):
    return export_all_outputs_controller(request, current_user["id"])


@router.get("/outputs/available", summary="Get available outputs for chaining")
def get_available_outputs(
    workspace_id: str,
    measurement_id: str,
    current_user: CurrentUser,
    accepted_types: str = None,
    accepted_plugin_ids: str = None,
):
    types_list = accepted_types.split(",") if accepted_types else None
    plugins_list = accepted_plugin_ids.split(",") if accepted_plugin_ids else None
    return get_available_outputs_controller(
        workspace_id, measurement_id, current_user["id"], types_list, plugins_list
    )


@router.get("", summary="Get all plugins for the current user")
def get_plugins(current_user: CurrentUser):
    return get_user_plugins_controller(current_user["id"])

@router.post("/validate", summary="Validate a plugin script without executing")
def validate_plugin(request: PluginValidate, _current_user: CurrentUser):
    return validate_plugin_controller(request)

@router.get("/{plugin_id}", summary="Get a specific plugin by id")
def get_plugin(plugin_id: str, current_user: CurrentUser):
    return get_plugin_by_id_controller(plugin_id, current_user["id"])


@router.post("", summary="Create a new plugin", status_code=201)
def create_plugin(request: PluginCreate, current_user: CurrentUser):
    return create_plugin_controller(request, current_user["id"])


@router.put("/{plugin_id}", summary="Update a plugin")
def update_plugin(plugin_id: str, request: PluginUpdate, current_user: CurrentUser):
    return update_plugin_controller(plugin_id, request, current_user["id"])


@router.patch("/{plugin_id}/toggle", summary="Toggle a plugin's active status")
def toggle_plugin(plugin_id: str, request: PluginToggle, current_user: CurrentUser):
    return toggle_plugin_controller(plugin_id, request, current_user["id"])


@router.delete("/{plugin_id}", summary="Delete a plugin")
def delete_plugin(plugin_id: str, current_user: CurrentUser):
    return delete_plugin_controller(plugin_id, current_user["id"])


@router.post("/{plugin_id}/execute", summary="Execute a plugin")
def execute_plugin(plugin_id: str, request: PluginExecute, current_user: CurrentUser):
    return execute_plugin_controller(plugin_id, request, current_user["id"])


@router.patch(
    "/{plugin_id}/update-from-marketplace",
    summary="Update installed plugin from marketplace",
)
def update_installed_plugin(plugin_id: str, current_user: CurrentUser):
    return update_installed_plugin_controller(plugin_id, current_user["id"])


@router.get(
    "/{plugin_id}/executions/latest",
    summary="Get the most recent execution for a measurement",
)
def get_latest_execution(
    plugin_id: str, workspace_id: str, measurement_id: str, current_user: CurrentUser
):
    return get_latest_execution_controller(
        plugin_id, workspace_id, measurement_id, current_user["id"]
    )