# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Full-SMS API
  description: Backend API for Single-Molecule Spectroscopy Analysis Web Platform
  version: 1.0.0
paths:
...
  /api/py/plugins:
    get:
      tags:
        - Plugins
      summary: Get all plugins for the current user
      operationId: get_plugins_api_py_plugins_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
        - HTTPBearer: []
    post:
      tags:
        - Plugins
      summary: Create a new plugin
      operationId: create_plugin_api_py_plugins_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
  /api/py/plugins/{plugin_id}:
    get:
      tags:
        - Plugins
      summary: Get a specific plugin by id
      operationId: get_plugin_api_py_plugins__plugin_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: plugin_id
          in: path
          required: true
          schema:
            type: string
            title: Plugin Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
        - Plugins
      summary: Update a plugin
      operationId: update_plugin_api_py_plugins__plugin_id__put
      security:
        - HTTPBearer: []
      parameters:
        - name: plugin_id
          in: path
          required: true
          schema:
            type: string
            title: Plugin Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Plugins
      summary: Delete a plugin
      operationId: delete_plugin_api_py_plugins__plugin_id__delete
      security:
        - HTTPBearer: []
      parameters:
        - name: plugin_id
          in: path
          required: true
          schema:
            type: string
            title: Plugin Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/py/plugins/{plugin_id}/toggle:
    patch:
      tags:
        - Plugins
      summary: Toggle a plugin's active status
      operationId: toggle_plugin_api_py_plugins__plugin_id__toggle_patch
      security:
        - HTTPBearer: []
      parameters:
        - name: plugin_id
          in: path
          required: true
          schema:
            type: string
            title: Plugin Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginToggle'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/py/plugins/{plugin_id}/execute:
    post:
      tags:
        - Plugins
      summary: Execute a plugin
      operationId: execute_plugin_api_py_plugins__plugin_id__execute_post
      security:
        - HTTPBearer: []
      parameters:
        - name: plugin_id
          in: path
          required: true
          schema:
            type: string
            title: Plugin Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginExecute'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/py/plugins/{plugin_id}/update-from-marketplace:
    patch:
      tags:
        - Plugins
      summary: Update installed plugin from marketplace
      operationId: >-
        update_installed_plugin_api_py_plugins__plugin_id__update_from_marketplace_patch
      security:
        - HTTPBearer: []
      parameters:
        - name: plugin_id
          in: path
          required: true
          schema:
            type: string
            title: Plugin Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MeasurementData:
      properties:
        microtimes:
          items:
            type: number
          type: array
          title: Microtimes
        abstimes:
          items:
            type: number
          type: array
          title: Abstimes
        channel:
          anyOf:
            - type: integer
            - type: 'null'
          title: Channel
        metadata:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Metadata
      type: object
      required:
        - microtimes
        - abstimes
      title: MeasurementData
...
    PluginConfig:
      properties:
        parameters:
          items:
            $ref: '#/components/schemas/PluginParameter'
          type: array
          title: Parameters
        outputs:
          items:
            $ref: '#/components/schemas/PluginOutput'
          type: array
          minItems: 1
          title: Outputs
        requiredPackages:
          items:
            type: string
          type: array
          title: Requiredpackages
      type: object
      required:
        - outputs
      title: PluginConfig
    PluginCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        description:
          anyOf:
            - type: string
              maxLength: 1000
            - type: 'null'
          title: Description
        version:
          type: string
          maxLength: 50
          title: Version
          default: 1.0.0
        config:
          $ref: '#/components/schemas/PluginConfig'
        script:
          type: string
          minLength: 1
          title: Script
      type: object
      required:
        - name
        - config
        - script
      title: PluginCreate
    PluginExecute:
      properties:
        workspace_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Workspace Id
        measurement_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Measurement Id
        parameters:
          additionalProperties: true
          type: object
          title: Parameters
        measurement_data:
          anyOf:
            - $ref: '#/components/schemas/MeasurementData'
            - type: 'null'
        upload_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Upload Id
      type: object
      title: PluginExecute
    PluginOutput:
      properties:
        id:
          type: string
          maxLength: 100
          minLength: 1
          title: Id
        label:
          type: string
          maxLength: 255
          minLength: 1
          title: Label
        type:
          type: string
          pattern: ^(plot|histogram|table|value|heatmap)$
          title: Type
      type: object
      required:
        - id
        - label
        - type
      title: PluginOutput
    PluginParameter:
      properties:
        id:
          type: string
          maxLength: 100
          minLength: 1
          title: Id
        label:
          type: string
          maxLength: 255
          minLength: 1
          title: Label
        type:
          type: string
          pattern: ^(number|text|select|checkbox|range)$
          title: Type
        default:
          anyOf:
            - {}
            - type: 'null'
          title: Default
        min:
          anyOf:
            - type: number
            - type: 'null'
          title: Min
        max:
          anyOf:
            - type: number
            - type: 'null'
          title: Max
        step:
          anyOf:
            - type: number
            - type: 'null'
          title: Step
        options:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Options
      type: object
      required:
        - id
        - label
        - type
      title: PluginParameter
    PluginToggle:
      properties:
        enabled:
          type: boolean
          title: Enabled
      type: object
      required:
        - enabled
      title: PluginToggle
    PluginUpdate:
      properties:
        name:
          anyOf:
            - type: string
              maxLength: 100
              minLength: 1
            - type: 'null'
          title: Name
        description:
          anyOf:
            - type: string
              maxLength: 1000
              minLength: 1
            - type: 'null'
          title: Description
        version:
          anyOf:
            - type: string
              maxLength: 50
            - type: 'null'
          title: Version
        config:
          anyOf:
            - $ref: '#/components/schemas/PluginConfig'
            - type: 'null'
        script:
          anyOf:
            - type: string
              minLength: 1
            - type: 'null'
          title: Script
      type: object
      title: PluginUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer